Move Reader to new reader module
